fixed middleend and backend, cam=n solve square equation

+++ headers/processor_h/stack.h
@@ -2,7 +2,7 @@
 #include <stdbool.h>
 #include <stdio.h>
 
-typedef int StackElem;
+typedef double StackElem;
 
 typedef struct 
 {
